+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 * Added builtin `device` column to csv-source. Automatically filters for devices that match the tag specified in this column, if present
 * Switched to using async endpoint
 
+### **Bug Fixes**
+* Better error-handling when an error is returned from the Virtual Device service
+
 ## 0.6.3
 ### **Enhancements**
 * Convenience methods for accessing outputFields and actualFields on the result object